forgot one file, d'oh!
git-svn-id: https://svn.code.sf.net/p/pinmame/code/trunk@4963 6f51dfdd-75df-4273-8f3b-818b1e6aa8bf

+++ src/driver.h
@@ -117,7 +117,7 @@ struct pinMachine {
   void (*reset)(void);
   void (*stop)(void);
 };
-extern void machine_add_timer(struct InternalMachineDriver *machine, void (*func)(int), int rate);
+extern void machine_add_timer(struct InternalMachineDriver *machine, void (*func)(int), double rate);
 #define DIAGLED_VERTICAL        0x100   /* Flag indicated DIAG LEDS are Vertically Positioned */
 #define MDRV_DIPS(no) \
   machine->pinmame.coreDips = (no);
